top 5 section

+++ database/migrations/2023_02_23_180614_create_titles_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->string('name');
             $table->enum('type', ['movie', 'serie']);
             $table->longText('overview');
-            $table->float('popularity');
+            $table->decimal('popularity');
             $table->timestamps();
         });
     }
